Added ability to have empty fields in form schemas
This allows us to use empty '{}' json objects in a form schema. The primary
reason for allowing this is to make using schemas in templates easier. So there
can always be a terminating comma in a list of fields which can then be
followed up by and empty json object so that the schema is still valid json (ie
all json objects are closed with a non-comma)